requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright==1.40.0
aiohttp>=3.9.0
gunicorn>=21.2.0
scikit-learn>=1.3.0
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.1.0
